Switched to the recommended poles and zeros for the Trillium Compact 120.
The poles and zeros in the Trillium Compact OBS users guide are an older
version and have since been refined. The new response includes additional
poles to better reflect the dropoff in the 1000Hz range. The change in the
passband is less than 0.1dB.

The new response does not reflect any changes in the actual instrument.
It is just a slight improvement in the accuracy of the transfer function
model.